Move input_overlay_auto_rotate_ out of retroarch.c and into input_driver.c
