+++ input/input_driver.h
@@ -23,7 +23,7 @@
 #include <sys/types.h>
 
 #ifdef HAVE_CONFIG_H
-#include "config.h"
+#include "../config.h"
 #endif /* HAVE_CONFIG_H */
 
 #include <boolean.h>
